fix: resolve security vulnerabilities and conversion warnings
Security fixes (found by static analysis):
- settingsserializer.cpp: fix off-by-one OOB write (>  vs >=  on groups.size())
- rawdatabaseimpl.cpp: null-check tox_pass_key_derive_with_salt() result before deref
- ipc.cpp: use explicit-length QString::fromUtf8/QByteArray on fixed-size IPC fields
  to prevent strlen reading past array bounds in shared memory
- core.cpp: replace unseeded rand() with std::mt19937 seeded from high_res_clock

Conversion warning fixes (-Wconversion / -Wsign-conversion):
- openal.cpp: add static_cast at 10 ALsizei/ALint/ALfloat/uint8_t conversion sites
- display.cpp: cast uint64_t to double before log() and division
- chatid.h: cast qHash return value to uint explicitly
- mockbootstraplistgenerator.h: cast getSelfUdpPort() to quint16

+++ src/core/chatid.h
@@ -38,7 +38,7 @@ protected:
 
 inline uint qHash(const ChatId& id)
 {
-    return qHash(id.getByteArray());
+    return static_cast<uint>(qHash(id.getByteArray()));
 }
 
 inline bool operator==(const std::reference_wrapper<const ChatId> lhs,
